+++ App/Context/AppContext.js
@@ -4,15 +4,20 @@ export const AppContext = createContext();
 
 const AppContextProvider = ({ children }) => {
     const [loggedInUser, setLoggedInUser] = useState(true)
+    const [transactions, setTransactions] = useState([])
+    const [incomeSubCatagories, setIncomeSubCatagories] = useState("")
+
+
+
     const [modal, setModal] = useState(false)
 
     return (
         <>
             <AppContext.Provider value={{
-                loggedInUser,
-                setLoggedInUser,
-                modal,
-                setModal
+                loggedInUser, setLoggedInUser,
+                transactions, setTransactions,
+                incomeSubCatagories, setIncomeSubCatagories,
+                modal, setModal
             }}>
                 {children}
             </AppContext.Provider>